refactor: Separate client logic into a new module
